[FE] Fix: Name change Changes to be committed: 	renamed:    src/api/index.ts -> src/api/getResourceWithAxios.ts 	modified:   src/api/student/getStudentDetailWithId.ts
